pyproject.toml
[project]
name = "endogenai-workflows"
version = "0.1.0"
description = "Endogenic development workflows, scripts, and agents for deterministic AI-assisted systems"
authors = [{name = "EndogenAI Contributors"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"

dependencies = [
    "watchdog>=3.0.0",
    "pyyaml>=6.0",
    "rank_bm25>=0.2",
    "textstat>=0.7.13",
    "markdown-it-py>=3.0.0",
    "networkx>=3.2.1",
    "mkdocs-material>=9.7.5",
    "youtube-transcript-api[dev]>=1.2.4",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "opentelemetry-semantic-conventions>=0.41b0",
    "requests>=2.31.0",
    "jsonschema>=4.26.0",
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--strict-markers --tb=short -v"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "io: marks tests that perform file I/O",
]

[tool.coverage.run]
source = ["scripts", "mcp_server"]
omit = ["*/tests/*", "**/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@abstractmethod",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

# Cookiecutter template — invocation reference:
# The dogma cookiecutter template lives at the repo root (cookiecutter.json).
# Scaffold a new dogma-based repo with:
#   uvx cookiecutter gh:EndogenAI/dogma        # from GitHub
#   uvx cookiecutter .                          # from a local clone
# No entry point is registered here because cookiecutter ships its own CLI.
# The template is referenced in docs/guides/product-fork-initialization.md.

# T2 (static linting) enforcement: Terminal file I/O redirection is monitored
# via pre-commit hook 'no-terminal-file-io-redirect' (see .pre-commit-config.yaml).
# Patterns flagged: > file, >> file, | tee file, | cat >> file
# Rationale: Shift file I/O from T1 (interactive terminal) to T2 (static tools)
# like create_file or replace_string_in_file. Reference: Issue #154, #151 research
